+++ variables.tf
@@ -62,7 +62,7 @@ variable "publisher_email" {
 }
 
 variable "additional_location" {
-  type        = list(string)
+  type        = list(map(string))
   description = "The name of the Azure Region in which the API Management Service should be expanded to."
   default     = []
 }