+++ CHANGELOG.md
@@ -4,6 +4,9 @@ Added
   * AZ-373: Add diagnostic settings, add NSG rule for APIM Management interface
   * AZ-475: Add log_destination_type variable
 
+Update
+  * AZ-409: Finalizing the Additional location feature (Added vNET configuration)
+
 # v3.2.0/v4.0.0 - 2020-12-11
 
 Updated